refactor: extract runtime diagnostic helpers
